Add files via upload
The MCMC-Global-ka and funcs_ub_globalfit are the original files from Alan - he has an updated version on his github that I do not have access to the funcs file for.

I have added the data from an iCCD measurement, however I think I need to be include a diffusion term, as the initial drop in the data is not replicable in the simulations. The beam from this measurement is Gaussian, so i think diffusion is still important.